Runner: turn runners into plugin implementations
The regular runner has been moved to its plugin module, and other
runner implementations should be based on the "plugin_interface"
definition only.  This can be seen on the remote runner, that
doesn't use any code from the regular runner.

The activation of plugins is now done through the dispatcher module,
in accordance with the other plugins.  The runner for a job can
now be chosen by the name of the runner plugin, instead of passing
the Python class itself there.  This is another improvements in
the Job API, as a job can now define something like:

   config = {
      'references': ['examples/tests/passtest.py:PassTest.test'],
      ...
      'test_runner': 'docker',
      ...
      }

And have the appropriate runner chosen by the job.  See
examples/jobs/podman.py for a working example.
